Clean up tax tests
